\name{OpenData}
\alias{OpenData}
\title{
Opens a variety of filetypes of data
}
\description{
Depending on file extension, chooses the appropriate function to open data. 
}
\usage{
OpenData(filename)
}
\arguments{
  \item{filename}{
a filename, with extension. In quotes.
}
}
\details{
}
\value{
The data set in the specified file. 
}
\references{
}
\author{
Amelia McNamara
}
\note{
}


\seealso{
 \code{\link{read.csv}},  \code{\link{load}},  \code{\link{dget}}
}
\examples{
}
\keyword{ ~data }
