\name{MobilizeSimple}
\alias{MobilizeSimple-package}
\alias{MobilizeSimple}
\docType{package}
\title{
Simple functions for Mobilize
}
\description{
Masks many complicated functions for creating maps and working with text into one-line functions. Written for the Mobilize project (\link{http://www.mobilizingcs.org})
}
t\details{
\tabular{ll}{
Package: \tab MobilizeSimple\cr
Type: \tab Package\cr
Version: \tab 1.2.6\cr
Date: \tab 2013-03-29\cr
License: \tab GPL\cr
}
}
\author{
Amelia McNamara
 <amelia.mcnamara@stat.ucla.edu>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
