\name{MobilizeSimple}
\alias{MobilizeSimple-package}
\alias{MobilizeSimple}
\docType{package}
\title{
Simple functions for Mobilize
}
\description{
Masks many complicated functions for creating maps and working with text into one-line functions. Written for the Mobilize project (\link{http://www.mobilizingcs.org})
}
\details{
\tabular{ll}{
Package: \tab MobilizeSimple\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-01-17\cr
License: \tab GPL\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Amelia McNamara
 <amelia.mcnamara@stat.ucla.edu>
}
\references{
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
